Pushed unmock up to helper.rb
Also switched osx group test to use quick_resource since it is cleaner.
