+++ test/helper.rb
@@ -151,6 +151,20 @@ class Minitest::Test
   def skip_windows!
     skip_until 2019, 12, 4, "These have never passed" if windows?
   end
+
+  def unmock(&blk)
+    # eg: resource = unmock { group "staff" }
+    require "fetchers/mock"
+    require "inspec/runner"
+
+    # TODO: there is WAY too much magic going on in here
+    runner = Inspec::Runner.new
+    runner.add_target("inspec.yml" => "name: inspec-shell")
+    profile = runner.target_profiles.first
+    ctx = profile.runner_context
+
+    ctx.load blk
+  end
 end
 
 class InspecTest < Minitest::Test
